06-generative-ai/03c-Agent-Tools-MCP/src/mcp_integration.py
"""
Simplified MCP Integration Example

This demonstrates the conceptual integration of MCP tools with AI agents.
In production, you would connect to actual MCP servers.
"""

import os
import json
from dotenv import load_dotenv
from typing import Dict, List, Any
from dataclasses import dataclass
from datetime import datetime

# Load environment variables
load_dotenv()

@dataclass
class MCPTool:
    """Represents an MCP tool configuration."""
    server_label: str
    server_url: str
    capabilities: List[str]
    description: str

class MCPToolRegistry:
    """Registry of available MCP tools."""
    
    def __init__(self):
        self.tools = {
            "documentation": MCPTool(
                server_label="microsoft-docs",
                server_url="https://docs.microsoft.com/mcp",
                capabilities=["search", "retrieve", "index"],
                description="Access Microsoft documentation"
            ),
            "code_samples": MCPTool(
                server_label="code-samples",
                server_url="https://samples.microsoft.com/mcp",
                capabilities=["search", "download", "validate"],
                description="Find and retrieve code samples"
            ),
            "api_reference": MCPTool(
                server_label="api-reference",
                server_url="https://api.microsoft.com/mcp",
                capabilities=["lookup", "schema", "examples"],
                description="API reference and schemas"
            ),
            "troubleshooting": MCPTool(
                server_label="troubleshooting",
                server_url="https://support.microsoft.com/mcp",
                capabilities=["diagnose", "solutions", "known_issues"],
                description="Troubleshooting guides and solutions"
            )
        }
    
    def get_tool(self, label: str) -> MCPTool:
        """Get MCP tool by label."""
        return self.tools.get(label)
    
    def list_tools(self) -> List[Dict]:
        """List all available MCP tools."""
        return [
            {
                "label": label,
                "description": tool.description,
                "capabilities": tool.capabilities
            }
            for label, tool in self.tools.items()
        ]

class MCPProtocolHandler:
    """Handles MCP protocol communication."""
    
    def __init__(self, tool: MCPTool):
        self.tool = tool
        self.session_id = None
        self.authenticated = False
    
    def connect(self) -> Dict:
        """Establish connection to MCP server."""
        # Simulated connection
        self.session_id = f"session_{datetime.now().strftime('%Y%m%d%H%M%S')}"
        self.authenticated = True
        
        return {
            "status": "connected",
            "session_id": self.session_id,
            "server": self.tool.server_label,
            "capabilities": self.tool.capabilities
        }
    
    def execute_command(self, command: str, params: Dict) -> Dict:
        """Execute MCP command."""
        if not self.authenticated:
            return {"error": "Not authenticated"}
        
        # Simulated command execution
        if command == "search":
            return self._search(params)
        elif command == "retrieve":
            return self._retrieve(params)
        elif command == "list_capabilities":
            return {"capabilities": self.tool.capabilities}
        else:
            return {"error": f"Unknown command: {command}"}
    
    def _search(self, params: Dict) -> Dict:
        """Simulated search operation."""
        query = params.get("query", "")
        
        # Simulated search results based on tool type
        if self.tool.server_label == "documentation":
            results = [
                {
                    "id": "doc1",
                    "title": "Getting Started with Azure AI",
                    "url": "https://docs.microsoft.com/azure-ai/getting-started",
                    "relevance": 0.95
                },
                {
                    "id": "doc2",
                    "title": "Azure AI Agent Service Overview",
                    "url": "https://docs.microsoft.com/azure-ai/agents",
                    "relevance": 0.88
                }
            ]
        elif self.tool.server_label == "code-samples":
            results = [
                {
                    "id": "sample1",
                    "title": "AI Agent with Custom Functions",
                    "language": "python",
                    "url": "https://samples.microsoft.com/ai-agent-functions",
                    "relevance": 0.92
                }
            ]
        else:
            results = []
        
        return {
            "query": query,
            "count": len(results),
            "results": results
        }
    
    def _retrieve(self, params: Dict) -> Dict:
        """Simulated retrieve operation."""
        resource_id = params.get("id", "")
        
        # Simulated content retrieval
        if resource_id == "doc1":
            content = """# Getting Started with Azure AI

Azure AI provides a comprehensive platform for building intelligent applications...
            
## Key Services
- Azure OpenAI Service
- Azure AI Search
- Azure AI Document Intelligence
            
## Quick Start
1. Create an Azure account
2. Provision AI services
3. Build your first application"""
        else:
            content = "Resource not found"
        
        return {
            "id": resource_id,
            "content": content,
            "metadata": {
                "last_updated": "2024-12-20",
                "author": "Microsoft",
                "version": "1.0"
            }
        }
    
    def disconnect(self) -> Dict:
        """Disconnect from MCP server."""
        self.authenticated = False
        self.session_id = None
        return {"status": "disconnected"}

def demonstrate_mcp_workflow():
    """Demonstrate MCP tool workflow."""
    print("=== MCP Tool Integration Demonstration ===\n")
    
    # Initialize registry
    registry = MCPToolRegistry()
    
    # List available tools
    print("Available MCP Tools:")
    for tool_info in registry.list_tools():
        print(f"- {tool_info['label']}: {tool_info['description']}")
        print(f"  Capabilities: {', '.join(tool_info['capabilities'])}")
    print()
    
    # Connect to documentation MCP tool
    doc_tool = registry.get_tool("documentation")
    handler = MCPProtocolHandler(doc_tool)
    
    # Establish connection
    print("Connecting to Documentation MCP Server...")
    connection = handler.connect()
    print(f"Status: {connection['status']}")
    print(f"Session: {connection['session_id']}")
    print()
    
    # Execute search
    print("Searching for 'Azure AI Agent'...")
    search_result = handler.execute_command("search", {"query": "Azure AI Agent"})
    print(f"Found {search_result['count']} results:")
    for result in search_result['results']:
        print(f"- {result['title']} (relevance: {result['relevance']})")
    print()
    
    # Retrieve document
    if search_result['results']:
        doc_id = search_result['results'][0]['id']
        print(f"Retrieving document '{doc_id}'...")
        doc_content = handler.execute_command("retrieve", {"id": doc_id})
        print("Content preview:")
        print(doc_content['content'][:200] + "...")
        print()
    
    # Disconnect
    print("Disconnecting...")
    handler.disconnect()
    print("MCP session closed.\n")

def create_mcp_enabled_agent_example():
    """Example of how to create an agent with MCP tools."""
    print("=== Creating MCP-Enabled Agent (Conceptual) ===\n")
    
    example_code = '''# Conceptual code for MCP-enabled agent
from azure.ai.agents import AgentsClient
from azure.ai.agents.models import McpTool

# Initialize MCP tools
doc_tool = McpTool(
    server_label="microsoft-docs",
    server_url="https://docs.microsoft.com/mcp"
)

code_tool = McpTool(
    server_label="code-samples",
    server_url="https://samples.microsoft.com/mcp"
)

# Create agent with MCP tools
agent = agents_client.create_agent(
    model="gpt-4",
    name="developer-assistant",
    instructions="""You are a developer assistant with access to:
    1. Microsoft documentation (real-time)
    2. Code samples repository
    3. API references
    
    Use these MCP tools to provide accurate, up-to-date information.""",
    tools=[doc_tool.definitions, code_tool.definitions]
)

# Agent can now use MCP tools automatically during conversations
'''
    
    print(example_code)
    print("\nKey Benefits of MCP Integration:")
    print("1. Real-time access to documentation")
    print("2. Always up-to-date information")
    print("3. Standardized tool interface")
    print("4. Scalable to multiple data sources")
    print("5. Secure, authenticated connections")

if __name__ == "__main__":
    # Run demonstration
    demonstrate_mcp_workflow()
    print("\n" + "="*60 + "\n")
    create_mcp_enabled_agent_example()