06-generative-ai/03c-Agent-Tools-MCP/requirements.txt
azure-identity
azure-ai-projects
azure-ai-agents>=1.0.0b1
python-dotenv
openai
requests